// green-go/src/shared/types.ts
export enum SelectedPage {
  Home = "home",
  AboutUs = "aboutus",
  Join = "join",
  OurClasses = "ourclasses",
  ContactUs = "contactus",
}

export interface BenefitType {
  icon: JSX.Element;
  title: string;
  description: string;
}

export interface ClassType {
  name: string;
  description?: string;
  image: string;
}